Revert "clang-tidy: use cmp functions"
This reverts commit 5a2469feb51dee9ca9b8a5d6a54b661e7fa736a7.

+++ src/image.cpp
@@ -332,7 +332,7 @@ void Image::printIFDStructure(BasicIo& io, std::ostream& out, Exiv2::PrintStruct
     }
 
     // Read the dictionary
-    for (size_t i = 0; i < dirLength; i++) {
+    for (int i = 0; i < dirLength; i++) {
       if (visits.contains(io.tell())) {  // #547
         throw Error(ErrorCode::kerCorruptedMetadata);
       }
